Split text fixtures (but I doubt it's enough to avoid Jasmine) Some of the "optimize" tests are too resource-intensive, Jasmine seems to run out of memory if it runs with the others, whereas they work fine on their own.
